Adding CC BY licence
file is in Markdown—can we automate conversion to adoc?

Does adding this alter anything about the machine-readable version attached to this repo? 

NOTE: we should definitely consider producing a Spanish-language version for the OpenRefine guide, and possibly maintaining a set of standard licence files in translation.

How do we square license v. licence? The latter is UN- and therefore GBIF-preferred style…